Adding mobility changes to ed attainment data
I altered Greg's mobility file to include the census fips code in his cleaned up CSV file, which allowed me to extract the maximum 7d rolling average change for each county and add it to the educational attainment csv. I used the fips code as a common key to merge these data sets. The output file ("VA_ed_data_cleaned" now includes a column for the maximum change in mobility for each county, and can be used for questions 3-4.